+++ src/main/java/tokyo/northside/omegat/textra/OmegatTextraMachineTranslation.java
@@ -172,9 +172,9 @@ public class OmegatTextraMachineTranslation extends BaseTranslate implements IMa
             throws Exception {
         if (enabled) {
             // Set TexTra access options
-            options.setLang(sLang.getLanguageCode(), tLang.getLanguageCode());
+            options.setLang(sLang, tLang);
             if (!options.isCombinationValid()) {
-                LOGGER.info(String.format("Invalid language combination for %s with source %s, "
+                LOGGER.info(String.format("Textra:Invalid language combination for %s with source %s, "
                                 + " and target %s.",
                         options.getModeName(), options.getSourceLang(), options.getTargetLang()));
                 return null;